+++ data_collector/file_system.js
@@ -1,0 +1,3 @@
+const secrets = require("./secrets");
+const fs = require("fs");
+
