+++ data_collector/test.js
@@ -1,0 +1,5 @@
+const test = require("ava");
+
+test("foo", t => {
+    t.pass();
+});
